[Fix] #15 - 폴더 변경, 이름 수정
